+++ pygce/models/bot.py
@@ -27,7 +27,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-from .garmin import GCDayTimeline
+from .garmin.timeline import GCDayTimeline
 
 
 class GarminConnectBot(object):